wrap navigator with redux
