+++ package.json
@@ -5,6 +5,9 @@
   "devDependencies": {
     "babel-eslint": "^8.0.1",
     "eslint": "^4.8.0",
+    "eslint-plugin-flowtype": "^2.37.0",
+    "eslint-plugin-prettier": "^2.3.1",
+    "eslint-plugin-react": "^7.4.0",
     "file-loader": "^0.11.2",
     "flow-bin": "^0.49.1",
     "jest-expo": "^21.0.2",
@@ -18,7 +21,8 @@
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
     "test": "node node_modules/jest/bin/jest.js --watch",
-    "flow": "flow"
+    "flow": "flow",
+    "lint": "node_modules/.bin/eslint -c .eslintrc . -f table"
   },
   "jest": {
     "preset": "jest-expo"
