+++ redux/selectors/navigation.js
@@ -1,0 +1,3 @@
+const getNav = ({ navigation }) => navigation;
+
+export { getNav };
